refactor: 코드 정리 및 토스 정상화
