+++ TechBlogNotificationsServer/src/main/java/com/dgh06175/techblognotificationsserver/TechBlogNotificationsServerApplication.java
@@ -5,8 +5,10 @@ import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
+@EnableJpaAuditing
 public class TechBlogNotificationsServerApplication {
 
     public static void main(String[] args) {
@@ -16,7 +18,7 @@ public class TechBlogNotificationsServerApplication {
     @Bean
     CommandLineRunner runScraper(ScrapController scrapController) {
         return args -> {
-            scrapController.run();
+            scrapController.scrapPosts();
         };
     }
 }
