+++ TechBlogNotificationsServer/src/main/java/com/dgh06175/techblognotificationsserver/controller/TechBlogController.java
@@ -3,6 +3,7 @@ package com.dgh06175.techblognotificationsserver.controller;
 import com.dgh06175.techblognotificationsserver.domain.Post;
 import com.dgh06175.techblognotificationsserver.repository.PostRepository;
 import java.util.List;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,15 +11,14 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
+@RequiredArgsConstructor
 public class TechBlogController {
 
-    @Autowired
-    private PostRepository postRepository;
+    private final PostRepository postRepository;
 
     @GetMapping("find-all")
     @ResponseBody
     public List<Post> findAll() {
-        List<Post> posts = postRepository.findAll();
-        return posts;
+        return postRepository.findAll();
     }
 }
