+++ TechBlogNotificationsServer/src/main/resources/application.yml
@@ -2,11 +2,7 @@ spring:
   application:
     name: TechBlogNotificationsServer
   jpa:
-    database-platform:
-      org.hibernate.dialect.H2Dialect
     hibernate:
       ddl-auto: update
-      properties:
-        hibernate.jdbc.time_zone: Asia/Seoul
   profiles:
     include: aws
